Fold in header. This can move out again if the program grows.
